requirements.txt
flask==2.2.5
flask-sqlalchemy==3.0.3
flask-login==0.6.2
flask-wtf==1.1.1
WTForms==3.0.1
email-validator==2.0.0
werkzeug==2.2.3
flask-bootstrap4==4.0.2
psycopg2-binary==2.9.10
